[SCRUM-95] refactor: 성공, 에러 메시지 디렉토리 위치 변경
