+++ src/main/java/com/projectlyrics/server/domain/artist/service/impl/ArtistCommandServiceImpl.java
@@ -8,7 +8,7 @@ import com.projectlyrics.server.domain.artist.repository.CommandArtistRepository
 import com.projectlyrics.server.domain.artist.repository.QueryArtistRepository;
 import com.projectlyrics.server.domain.artist.service.ArtistCommandService;
 import com.projectlyrics.server.global.exception.FeelinException;
-import com.projectlyrics.server.global.message.ErrorCode;
+import com.projectlyrics.server.domain.common.message.ErrorCode;
 import java.time.Clock;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
