+++ src/main/java/com/projectlyrics/server/global/exception/NotFoundException.java
@@ -1,6 +1,6 @@
 package com.projectlyrics.server.global.exception;
 
-import com.projectlyrics.server.global.message.ErrorCode;
+import com.projectlyrics.server.domain.common.message.ErrorCode;
 
 public class NotFoundException extends FeelinException {
